# Move initialization of refence variables infront of `for` loops