In my opinion, the variables should be initialized before the `for` loops and freed after. However, I'm not sure whether this is important at all. Does Julia optimize this automatically?
I added a small benchmarking below. It doesn't really get faster, but the allocations decrease.
Any opinions on that?

Benchmarking of `count_interfaces` on my local machine and on 1 rank:
Before:
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):  397.887 μs …   5.934 ms  ┊ GC (min … max): 0.00% … 92.39%
 Time  (median):     405.502 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   421.143 μs ± 266.739 μs  ┊ GC (mean ± σ):  3.27% ±  4.78%

         ▃▇█▆▅▆▃▂
  ▁▂▃▆▆▆▆█████████▇▆▅▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  398 μs           Histogram: frequency by time          437 μs <

 Memory estimate: 211.38 KiB, allocs estimate: 6782.
```

After:
```
julia> @benchmark Trixi.count_interfaces(forest, ndims)
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):  411.872 μs …   7.560 ms  ┊ GC (min … max): 0.00% … 93.82%
 Time  (median):     418.337 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   432.496 μs ± 279.452 μs  ┊ GC (mean ± σ):  2.87% ±  4.16%

    ▁▃▅▆▇█████▇▆▅▅▄▃▃▂▃▂▂▁▁▁▁▁ ▁                                ▃
  ▂▅█████████████████████████████▇█▇██▇▇██▇█▇▇▇█▇█▇▇▆▆▆▆▆▇▆▇▅▆▆ █
  412 μs        Histogram: log(frequency) by time        451 μs <

 Memory estimate: 158.02 KiB, allocs estimate: 3367.
```

And within a simulation:
```
julia> @benchmark Trixi.count_interfaces(forest, ndims)
```
Before:
```
──────────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                       Time                    Allocations
                                    ───────────────────────   ────────────────────────
         Tot / % measured:               106ms /  98.0%           27.9MiB /  99.7%

Section                     ncalls     time    %tot     avg     alloc    %tot      avg
──────────────────────────────────────────────────────────────────────────────────────
AMR                             36   56.4ms   54.1%  1.57ms   25.7MiB   92.3%   731KiB
  adapt                         36   55.8ms   53.5%  1.55ms   25.6MiB   92.0%   729KiB
    solver                      36   36.8ms   35.3%  1.02ms   25.3MiB   90.8%   719KiB
    mesh                        36   19.0ms   18.2%   527μs    351KiB    1.2%  9.74KiB
    ~adapt~                     36   49.3μs    0.0%  1.37μs   5.62KiB    0.0%     160B
  indicator                     36    432μs    0.4%  12.0μs   8.61KiB    0.0%     245B
  ~AMR~                         36    197μs    0.2%  5.47μs   70.6KiB    0.2%  1.96KiB
rhs!                           926   38.8ms   37.2%  41.9μs   9.33KiB    0.0%    10.3B
  volume integral              926   18.3ms   17.6%  19.8μs     0.00B    0.0%    0.00B
  interface flux               926   6.64ms    6.4%  7.17μs     0.00B    0.0%    0.00B
  boundary flux                926   2.71ms    2.6%  2.92μs     0.00B    0.0%    0.00B
  prolong2interfaces           926   2.52ms    2.4%  2.72μs     0.00B    0.0%    0.00B
  mortar flux                  926   2.08ms    2.0%  2.24μs     0.00B    0.0%    0.00B
  prolong2mortars              926   1.93ms    1.8%  2.08μs     0.00B    0.0%    0.00B
  Jacobian                     926   1.75ms    1.7%  1.89μs     0.00B    0.0%    0.00B
  surface integral             926   1.57ms    1.5%  1.69μs     0.00B    0.0%    0.00B
  ~rhs!~                       926    726μs    0.7%   784ns   9.33KiB    0.0%    10.3B
  reset ∂u/∂t                  926    316μs    0.3%   341ns     0.00B    0.0%    0.00B
  prolong2boundaries           926    209μs    0.2%   226ns     0.00B    0.0%    0.00B
  source terms                 926   17.5μs    0.0%  18.9ns     0.00B    0.0%    0.00B
I/O                              7   3.84ms    3.7%   548μs    303KiB    1.1%  43.2KiB
  ~I/O~                          7   2.80ms    2.7%   400μs    110KiB    0.4%  15.7KiB
  save solution                  3   1.01ms    1.0%   337μs    192KiB    0.7%  64.1KiB
  get element variables          3   26.0μs    0.0%  8.66μs      816B    0.0%     272B
  save mesh                      3    346ns    0.0%   115ns     0.00B    0.0%    0.00B
  get node variables             3    246ns    0.0%  82.0ns     0.00B    0.0%    0.00B
initial condition AMR            1   2.99ms    2.9%  2.99ms   1.68MiB    6.0%  1.68MiB
  AMR                            3   2.91ms    2.8%   969μs   1.68MiB    6.0%   573KiB
    adapt                        3   2.86ms    2.7%   952μs   1.58MiB    5.7%   539KiB
      solver                     2   1.98ms    1.9%   988μs   1.55MiB    5.6%   795KiB
      mesh                       3    876μs    0.8%   292μs   25.8KiB    0.1%  8.59KiB
      ~adapt~                    3   4.44μs    0.0%  1.48μs   1.81KiB    0.0%     619B
    indicator                    3   29.3μs    0.0%  9.75μs   6.48KiB    0.0%  2.16KiB
    ~AMR~                        3   21.0μs    0.0%  6.99μs   95.1KiB    0.3%  31.7KiB
  ~initial condition AMR~        1   85.1μs    0.1%  85.1μs      752B    0.0%     752B
analyze solution                 3   1.40ms    1.3%   465μs    155KiB    0.5%  51.8KiB
calculate dt                   186    786μs    0.8%  4.22μs     0.00B    0.0%    0.00B
──────────────────────────────────────────────────────────────────────────────────────
```

After:
```
──────────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                       Time                    Allocations      
                                    ───────────────────────   ────────────────────────
         Tot / % measured:               105ms /  97.9%           23.6MiB /  99.6%    

Section                     ncalls     time    %tot     avg     alloc    %tot      avg
──────────────────────────────────────────────────────────────────────────────────────
AMR                             36   55.3ms   53.6%  1.54ms   21.6MiB   91.8%   614KiB
  adapt                         36   54.7ms   53.0%  1.52ms   21.5MiB   91.5%   611KiB
    solver                      36   36.1ms   34.9%  1.00ms   21.1MiB   90.0%   602KiB
    mesh                        36   18.5ms   18.0%   515μs    351KiB    1.5%  9.74KiB
    ~adapt~                     36   44.0μs    0.0%  1.22μs   5.62KiB    0.0%     160B
  indicator                     36    434μs    0.4%  12.1μs   8.61KiB    0.0%     245B
  ~AMR~                         36    198μs    0.2%  5.50μs   70.6KiB    0.3%  1.96KiB
rhs!                           926   38.9ms   37.7%  42.0μs   9.33KiB    0.0%    10.3B
  volume integral              926   18.4ms   17.8%  19.8μs     0.00B    0.0%    0.00B
  interface flux               926   6.68ms    6.5%  7.22μs     0.00B    0.0%    0.00B
  boundary flux                926   2.72ms    2.6%  2.94μs     0.00B    0.0%    0.00B
  prolong2interfaces           926   2.54ms    2.5%  2.74μs     0.00B    0.0%    0.00B
  mortar flux                  926   2.09ms    2.0%  2.26μs     0.00B    0.0%    0.00B
  prolong2mortars              926   1.92ms    1.9%  2.07μs     0.00B    0.0%    0.00B
  Jacobian                     926   1.74ms    1.7%  1.88μs     0.00B    0.0%    0.00B
  surface integral             926   1.55ms    1.5%  1.68μs     0.00B    0.0%    0.00B
  ~rhs!~                       926    692μs    0.7%   748ns   9.33KiB    0.0%    10.3B
  reset ∂u/∂t                  926    321μs    0.3%   346ns     0.00B    0.0%    0.00B
  prolong2boundaries           926    209μs    0.2%   226ns     0.00B    0.0%    0.00B
  source terms                 926   17.7μs    0.0%  19.1ns     0.00B    0.0%    0.00B
I/O                              7   4.09ms    4.0%   585μs    303KiB    1.3%  43.2KiB
  ~I/O~                          7   2.96ms    2.9%   423μs    110KiB    0.5%  15.7KiB
  save solution                  3   1.10ms    1.1%   367μs    192KiB    0.8%  64.1KiB
  get element variables          3   26.9μs    0.0%  8.97μs      816B    0.0%     272B
  get node variables             3    374ns    0.0%   125ns     0.00B    0.0%    0.00B
  save mesh                      3    180ns    0.0%  60.0ns     0.00B    0.0%    0.00B
initial condition AMR            1   2.87ms    2.8%  2.87ms   1.47MiB    6.3%  1.47MiB
  AMR                            3   2.79ms    2.7%   929μs   1.47MiB    6.2%   501KiB
    adapt                        3   2.74ms    2.7%   913μs   1.37MiB    5.8%   467KiB
      solver                     2   1.90ms    1.8%   951μs   1.34MiB    5.7%   687KiB
      mesh                       3    833μs    0.8%   278μs   25.8KiB    0.1%  8.59KiB
      ~adapt~                    3   3.20μs    0.0%  1.07μs   1.81KiB    0.0%     619B
    indicator                    3   28.9μs    0.0%  9.65μs   6.48KiB    0.0%  2.16KiB
    ~AMR~                        3   20.5μs    0.0%  6.83μs   95.1KiB    0.4%  31.7KiB
  ~initial condition AMR~        1   84.9μs    0.1%  84.9μs      752B    0.0%     752B
analyze solution                 3   1.27ms    1.2%   424μs    155KiB    0.6%  51.8KiB
calculate dt                   186    806μs    0.8%  4.34μs     0.00B    0.0%    0.00B
──────────────────────────────────────────────────────────────────────────────────────
```