@bennibolm 2025-02-03
> This makes the code more non-local and a bit harder to read. Is the gain worth this? (This is an honest question, not a rhetorical one)

That's why I wanted to talk about it. On my opinion, we can keep it as it is since it doesn't really speed up the simulation.
I was just a little bit confused about decreasing allocations, but no influence on the time. Is there an easy explanation for this? Does Julia somehow optimize it automatically?